data_collection/requirements.txt
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
emoji==1.2.0
idna==2.10
isort==5.7.0
joblib==1.0.1
json-lines==0.5.0
kiwisolver==1.3.1
langid==1.1.6
lazy-object-proxy==1.4.3
matplotlib==3.4.2
mccabe==0.6.1
mplcairo==0.4
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.1
Pillow==8.2.0
praw==7.1.0
prawcore==1.5.0
pybind11==2.6.2
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.15.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.61.0
tweepy==3.10.0
tweet-preprocessor==0.6.0
update-checker==0.18.0
urllib3==1.26.3
webscrape-master @ file:///C:/Users/eirik/PycharmProjects-38/Master
websocket-client==0.57.0
wrapt==1.12.1
